refactor: рефакторинг + условный рендер страницы
