+++ src/components/app/app.jsx
@@ -1,32 +1,22 @@
-import { useEffect } from "react";
-import { fetchData } from "../../services/reducers/get-data";
-import { DndProvider } from 'react-dnd'
-import { HTML5Backend } from 'react-dnd-html5-backend';
-import { useDispatch, useSelector } from "react-redux";
-import AppHeader from "../app-header/app-header";
-import styles from "./app.module.css";
-import BurgerConstructor from "../burger-constructor/burger-constructor";
-import BurgerIngredients from "../burger-ingredients/burger-ingredients";
-
+import { BrowserRouter, Routes, Route, Navigate} from "react-router-dom";
+import { useSelector } from "react-redux";
+import Main from '../main/main'
+import { LoginPage, RegisterPage, ForgotPasswordPage, ProfilePage, IngridientsPage, ResetPasswordPage, PageNotFoundPage} from "../../pages";
 function App() {
-  const dispatch = useDispatch() 
-  useEffect(()=>{
-    dispatch(fetchData())
-  },[dispatch])
-  const ingridients = useSelector(state=>state.getData)
+  const isNavigate = useSelector((state) => state.resetData.status)
   return (
-    <div className="grid justify-items-center">
-      <header className={styles.header}>
-        <AppHeader />
-      </header>
-      <DndProvider backend={HTML5Backend}>
-        {ingridients && <main className="grid grid-cols-2 gap-16 pt-20 absolute">
-          <BurgerIngredients/>
-          <BurgerConstructor/>
-        </main>}
-      </DndProvider>
-    </div>
+      <BrowserRouter>
+        <Routes>
+          <Route path="/" element={<Main />} />
+          <Route path="/login" element={<LoginPage />} />
+          <Route path="/register" element={<RegisterPage />} />
+          <Route path="/forgot-password" element={isNavigate ? (<Navigate replace to="/reset-password" />):(<ForgotPasswordPage />)} />
+          <Route path="/reset-password" element={<ResetPasswordPage />} />
+          <Route path="/profile" element={<ProfilePage />} />
+          <Route path="/ingredients/:id" element={<IngridientsPage />} />
+          <Route path="*" element={<PageNotFoundPage />} />
+        </Routes>
+      </BrowserRouter>
   );
 }
-
 export default App;
