--- src/components/app/app.module.css
@@ -1,5 +1,0 @@
-.header{
-  width: 100%;
-  height: 88px;
-  background-color: #1C1C21
-}
