+++ src/index.jsx
@@ -2,28 +2,14 @@ import React from "react";
 import ReactDOM from "react-dom";
 import { store } from "./services/store";
 import { Provider } from "react-redux";
-import { BrowserRouter, Routes, Route } from "react-router-dom";
-import { LoginPage, RegisterPage, ForgotPasswordPage, ProfilePage, IngridientsPage, ResetPasswordPage, PageNotFoundPage} from "./pages";
-
+import App from "./components/app/app";
 import "./index.css";
 
-import App from "./components/app/app";
 
 ReactDOM.render(
   <React.StrictMode>
     <Provider store={store}>
-      <BrowserRouter>
-        <Routes>
-          <Route path="/" element={<App />} />
-          <Route path="/login" element={<LoginPage />} />
-          <Route path="/register" element={<RegisterPage />} />
-          <Route path="/forgot-password" element={<ForgotPasswordPage />} />
-          <Route path="/reset-password" element={<ResetPasswordPage />} />
-          <Route path="/profile" element={<ProfilePage />} />
-          <Route path="/ingredients/:id" element={<IngridientsPage />} />
-          <Route path="*" element={<PageNotFoundPage />} />
-        </Routes>
-      </BrowserRouter>
+      <App/>
     </Provider>
   </React.StrictMode>,
   document.getElementById("root")
